test: refactor http response code to response util
