+++ src/test/java/nextstep/subway/acceptance/LineAcceptanceTest.java
@@ -12,7 +12,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import static nextstep.subway.acceptance.LineSteps.*;
-import static nextstep.subway.utils.HTTPResponseCode.응답의_STATUS_검증;
+import static nextstep.subway.utils.ResponseUtils.응답의_STATUS_검증;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("지하철 노선 관리 기능")
