+++ src/test/java/nextstep/subway/utils/ResponseUtils.java
@@ -6,8 +6,14 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.springframework.http.HttpStatus;
 
-public class HTTPResponseCode {
+public class ResponseUtils {
+
     public static void 응답의_STATUS_검증(ExtractableResponse<Response> response, HttpStatus status) {
         assertThat(response.statusCode()).isEqualTo(status.value());
     }
+
+    public static long 응답에서_id_조회(ExtractableResponse<Response> response) {
+        return response.jsonPath().getLong("id");
+    }
 }
+
